Icicles and Moving Platforms
Our Firewall was just a bunch of Icicles stacked on top of each other, so I changed that and made them separate prefabs.
Fixed the Icicle prefab so that it will fall and then kill the player on hit.
Modified rendering of FireBlock and Iceblock sprites.
Deleted old Ice and Fire blocks
Created Ice and Fire Tilesets
Created Moving Platform and Oscillate Script.